add golangci.yml, add docs, refactor
