+++ .golangci.yml
@@ -1,0 +1,29 @@
+# Refer to golangci-lint's example config file for more options and information:
+# https://github.com/golangci/golangci-lint/blob/master/.golangci.example.yml
+
+run:
+  timeout: 5m
+  modules-download-mode: readonly
+
+# deadcode errcheck
+# gosimple govet
+# ineffassign staticcheck
+# typecheck unused varcheck
+linters:
+  enable:
+    - deadcode
+    - errcheck
+    - gosimple
+    - revive
+    - govet
+    - ineffassign
+    - staticcheck
+    - typecheck
+    - unused
+    - varcheck
+    - nilerr
+
+issues:
+  exclude-use-default: false
+  max-issues-per-linter: 0
+  max-same-issues: 0
