+++ main.go
@@ -5,7 +5,7 @@ import (
 	"crypto/md5"
 	"encoding/json"
 	"fmt"
-	"golang-developer-test-task/infrastructure/redis_db"
+	redclient2 "golang-developer-test-task/redclient"
 	"golang-developer-test-task/structs"
 	"html/template"
 	"io"
@@ -17,81 +17,13 @@ import (
 	"github.com/go-redis/redis/v9"
 	"github.com/mailru/easyjson"
 	"go.uber.org/zap"
-	"golang.org/x/text/encoding/charmap"
 )
 
-// https://stackoverflow.com/questions/11692860/how-can-i-efficiently-download-a-large-file-using-go
-type (
-	JsonObjectsProcessorFunc func(io.Reader) error
-
-	DBProcessor struct {
-		client *redis_db.RedisClient
-		logger *zap.Logger
-	}
-)
-
-func (f *DBProcessor) ProcessJSONs(reader io.Reader) (err error) {
-	bs, err := io.ReadAll(reader)
-	if err != nil {
-		f.logger.Error("error inside ProcessJSONs during ReadAll",
-			zap.Error(err))
-		return err
-	}
-	dec := charmap.Windows1251.NewDecoder()
-	out, err := dec.Bytes(bs)
-	var infoList structs.InfoList
-	err = easyjson.Unmarshal(out, &infoList)
-	if err != nil {
-		f.logger.Error("error inside ProcessJSONs during Unmarshal",
-			zap.Error(err))
-		return err
-	}
-	//fmt.Println("INFOLIST", infoList)
-	//fmt.Println()
-	for _, info := range infoList {
-		// TODO: should we accumulate json objects to insert?
-		//  or restrict number of goroutines?
-		go func(info structs.Info) {
-			err = f.client.AddValue(context.Background(), info)
-			if err != nil {
-				f.logger.Error("error inside ProcessJSONs' goroutine during adding value",
-					zap.Error(err))
-				return
-			}
-		}(info)
-	}
-	return nil
-}
-
-func (f *DBProcessor) ProcessFileFromURL(url string, processor JsonObjectsProcessorFunc) (err error) {
-	resp, err := http.Get(url)
-	if err != nil {
-		f.logger.Error("error inside ProcessFileFromURL",
-			zap.Error(err))
-		return err
-	}
-	defer resp.Body.Close()
-	err = processor(resp.Body)
-	return err
-}
-
-func (f *DBProcessor) ProcessFileFromRequest(r *http.Request, fileName string, processor JsonObjectsProcessorFunc) (err error) {
-	file, _, err := r.FormFile(fileName)
-	if err != nil {
-		f.logger.Error("error inside ProcessFileFromRequest",
-			zap.Error(err))
-		return err
-	}
-	defer file.Close()
-	err = processor(file)
-	return err
-}
-
-type URLObject struct {
+type urlObject struct {
 	URL string `json:"url"`
 }
 
-type SearchObject struct {
+type searchObject struct {
 	GlobalID       *int    `json:"global_id,omitempty"`
 	SystemObjectID *string `json:"system_object_id,omitempty"`
 	ID             *int    `json:"id,omitempty"`
@@ -101,7 +33,7 @@ type SearchObject struct {
 	Offset         int     `json:"offset"`
 }
 
-type PaginationObject struct {
+type paginationObject struct {
 	Size        int              `json:"size"`
 	Offset      int              `json:"offset"`
 	HasNext     bool             `json:"hasNext"`
@@ -122,11 +54,16 @@ func main() {
 	}()
 
 	ctx := context.TODO()
-	conf := redis_db.RedisConfig{}
+	conf := redclient2.RedisConfig{}
 	conf.Load()
 
-	client := redis_db.RedisConnect(ctx, conf)
-	defer client.Close()
+	client := redclient2.NewRedisClient(ctx, conf)
+	defer func() {
+		err = client.Close()
+		if err != nil {
+			panic(err)
+		}
+	}()
 
 	dbLogic := DBProcessor{client: client, logger: logger}
 	mux := http.NewServeMux()
@@ -154,8 +91,7 @@ func main() {
 			w.WriteHeader(http.StatusInternalServerError)
 			return
 		}
-		defer r.Body.Close()
-		var urlObj URLObject
+		var urlObj urlObject
 		err = json.Unmarshal(bs, &urlObj)
 		if err != nil {
 			w.WriteHeader(http.StatusInternalServerError)
@@ -181,7 +117,7 @@ func main() {
 		//	return
 		//}
 		//defer r.Body.Close()
-		var searchObj SearchObject
+		var searchObj searchObject
 		//err = json.Unmarshal(bs1, &searchObj)
 		//if err != nil {
 		//	w.WriteHeader(http.StatusInternalServerError)
@@ -215,7 +151,7 @@ func main() {
 		}
 
 		var v string
-		paginationObj := PaginationObject{}
+		paginationObj := paginationObject{}
 		paginationObj.Data = make(structs.InfoList, 0)
 		if !multiple {
 			v, err = client.Get(ctx, searchStr).Result()
