+++ redclient/config.go
@@ -1,13 +1,13 @@
-package redis_db
+package redclient
 
-//var Config RedisConfig
-
+// RedisConfig is struct for storing data about path to Redis Storage
 type RedisConfig struct {
 	Addr     string
 	Password string
 	DB       int
 }
 
+// Load is useful for loading RedisConfig data
 func (r *RedisConfig) Load() {
 	r.Addr = "localhost:6379"
 	r.Password = ""
